+++ src/app/api/types.ts
@@ -1,0 +1,6 @@
+export type PatchItemPayload = Partial<{
+  name: string;
+  memo: string;
+  imageUrl: string;
+  isCompleted: boolean;
+}>;
