+++ src/app/components/layout/Navbar.tsx
@@ -3,7 +3,8 @@
 import Image from "next/image";
 import logoLarge from "@/app/ui/image/logo/doit-logo-large.png";
 import logoSmall from "@/app/ui/image/logo/doit-logo-small.png";
-import useInnerWidth from "../hook/getWindowSize";
+import useInnerWidth from "../../hooks/getWindowSize";
+import Link from "next/link";
 
 const Navbar = () => {
   const isMobile = useInnerWidth() < 768;
@@ -16,14 +17,16 @@ const Navbar = () => {
       md:px-6
       lg:px-[360px]"
     >
-      <Image
-        src={isMobile ? logoSmall : logoLarge}
-        alt="사이트 이름 do it이 적혀있는 로고"
-        width={imgWidth}
-        height={imgHeight}
-        style={{ width: imgWidth, height: imgHeight }}
-        priority
-      />
+      <Link href="/" className="cursor-pointer">
+        <Image
+          src={isMobile ? logoSmall : logoLarge}
+          alt="사이트 이름 do it이 적혀있는 로고"
+          width={imgWidth}
+          height={imgHeight}
+          style={{ width: imgWidth, height: imgHeight }}
+          priority
+        />
+      </Link>
     </nav>
   );
 };
