+++ src/app/constants/form.ts
@@ -1,0 +1,10 @@
+import { FormInputConfig } from "../types/form";
+
+export const TODO_FORM_INPUTS: FormInputConfig[] = [
+  {
+    id: 1,
+    type: "text",
+    name: "todo",
+    placeholder: "할 일을 입력해주세요.",
+  },
+];
