+++ src/app/page.tsx
@@ -2,71 +2,17 @@
 
 import { FormEvent, useState, useCallback } from "react";
 import * as z from "zod/v4";
-import Navbar from "./components/Navbar";
-import FormInput from "./components/FormInput";
-import TodoList from "./components/TodoList";
+import Navbar from "./components/layout/Navbar";
+import FormInput from "./components/common/FormInput";
+import TodoList from "./components/todo/TodoList";
 import { todoSchema } from "./schemas/todo";
-import { addTodo } from "./api/post/addTodo";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-
-const inputs = [
-  {
-    id: 1,
-    type: "text",
-    name: "todo",
-    placeholder: "할 일을 입력해주세요.",
-  },
-];
+import { TODO_FORM_INPUTS } from "./constants/form";
+import { useAddTodo } from "./hooks/useTodoMutations";
 
 const HomePage = () => {
-  const queryClient = useQueryClient();
   const [errors, setErrors] = useState<Record<string, string>>({});
 
-  // Todo 타입 정의
-  type Todo = {
-    id: number;
-    name: string;
-    isCompleted: boolean;
-  };
-
-  // 뮤테이션 컨텍스트 타입 정의
-  type MutationContext = {
-    previousTodos?: Todo[];
-  };
-
-  const addTodoMutation = useMutation<void, Error, string, MutationContext>({
-    mutationFn: (name: string) => addTodo(name),
-    onMutate: async (newTodo): Promise<MutationContext> => {
-      // 낙관적 업데이트
-      await queryClient.cancelQueries({ queryKey: ["todos"] });
-
-      const previousTodos = queryClient.getQueryData<Todo[]>(["todos"]);
-
-      queryClient.setQueryData<Todo[]>(["todos"], (old) => {
-        if (!old) return [];
-        return [
-          ...old,
-          {
-            id: Date.now(), // 임시 ID
-            name: newTodo,
-            isCompleted: false,
-          },
-        ];
-      });
-
-      return { previousTodos };
-    },
-    onError: (err, variables, context) => {
-      // 에러 발생 시 이전 상태로 롤백
-      if (context?.previousTodos) {
-        queryClient.setQueryData(["todos"], context.previousTodos);
-      }
-    },
-    onSettled: () => {
-      // 최종적으로 서버 데이터와 동기화
-      queryClient.invalidateQueries({ queryKey: ["todos"] });
-    },
-  });
+  const addTodoMutation = useAddTodo();
 
   const handleSubmit = useCallback(
     (e: FormEvent<HTMLFormElement>) => {
@@ -103,7 +49,7 @@ const HomePage = () => {
       <Navbar />
       <main className="px-4 pb-4 md:px-6 lg:px-[360px]">
         <FormInput
-          inputs={inputs}
+          inputs={TODO_FORM_INPUTS}
           handleSubmit={handleSubmit}
           containerStyle="flex gap-4"
           errors={errors}
