+++ src/app/types/form.ts
@@ -1,0 +1,7 @@
+// types/form.ts
+export interface FormInputConfig {
+  id: number;
+  type: string;
+  name: string;
+  placeholder: string;
+}
