+++ src/app/types/todo.ts
@@ -1,0 +1,14 @@
+// types/todo.ts
+export interface Todo {
+  id: number;
+  name: string;
+  isCompleted: boolean;
+}
+
+export interface TodoFormData {
+  todo: string;
+}
+
+export interface MutationContext {
+  previousTodos?: Todo[];
+}
